refactor(wasm): split api/gpu_meshes.rs into prepass + batch modules
gpu_meshes.rs (1441 LOC) split into a gpu_meshes/ directory: prepass.rs
(build_pre_pass_once / build_pre_pass_streaming + job emission + fold_u128_to_u32
+ affinity tests) and batch.rs (produce_batch + the 3 process_geometry_batch*
wrappers + PartitionedBatch + ElementMeshOutput), behind mod.rs.

Pure code motion. The only edits beyond moving code are the 53 `super::X`
references that shift when the file nests a level deeper: rewritten to the
absolute `crate::api::X` spelling (compiler-validated; they point at the same
items). All #[wasm_bindgen] attributes/methods/signatures preserved.

The regenerated packages/wasm/pkg/ifc-lite.d.ts is included: the split groups
the two buildPrePass* methods into one impl block instead of interleaved, so
wasm-bindgen emits them in a different order. Proven to be a pure reorder, the
sorted member set is byte-identical (no method/property/signature added,
removed, or changed). Warning set identical to baseline; cargo test
-p ifc-lite-wasm green (12); wasm32 build clean; build-wasm.sh succeeds.
